+++ src/main/kotlin/io/runescript/plugin/ide/completion/RsCompletionProviderBase.kt
@@ -1,0 +1,6 @@
+package io.runescript.plugin.ide.completion
+
+import com.intellij.codeInsight.completion.CompletionParameters
+import com.intellij.codeInsight.completion.CompletionProvider
+
+abstract class RsCompletionProviderBase : CompletionProvider<CompletionParameters>()