+++ src/main/kotlin/io/runescript/plugin/ide/completion/RsCompletionProviderProc.kt
@@ -1,0 +1,7 @@
+package io.runescript.plugin.ide.completion
+
+import io.runescript.plugin.ide.completion.insertHandler.RsScriptInsertHandler
+import io.runescript.plugin.lang.stubs.index.RsProcScriptIndex
+
+class RsCompletionProviderProc() :
+    RsCompletionProviderScriptBase(RsProcScriptIndex.KEY, RsScriptInsertHandler("~"))